backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv==1.0.0
google-generativeai
google-cloud-aiplatform
requests
elevenlabs
perplexityai
google-adk==1.3.0
PyPDF2
markdown2
reportlab
